refactor: extract inject decorator function
